CMA-19 Add messages functionality

+++ server/src/main/java/com/project/cisco/service/impl/LanguageServiceImpl.java
@@ -2,8 +2,10 @@ package com.project.cisco.service.impl;
 
 import com.project.cisco.database.entity.Language;
 import com.project.cisco.database.repository.LanguageRepository;
+import com.project.cisco.database.repository.MessageRepository;
 import com.project.cisco.dto.LanguageDto;
 import com.project.cisco.exception.LengthConstraintViolationException;
+import com.project.cisco.exception.NotAllowedLanguageException;
 import com.project.cisco.exception.NotFoundException;
 import com.project.cisco.exception.UniqueConstraintViolationException;
 import com.project.cisco.mapper.LanguageMapper;
@@ -20,6 +22,8 @@ import java.util.Optional;
 public class LanguageServiceImpl implements LanguageService {
     @Autowired
     private LanguageRepository languageRepository;
+    @Autowired
+    private MessageRepository messageRepository;
     @Autowired
     private LanguageMapper languageMapper;
 
@@ -57,6 +61,14 @@ public class LanguageServiceImpl implements LanguageService {
         if (languageOptional.isEmpty()) {
             throw new NotFoundException("Language with given id does not exist");
         }
+        if(languageOptional.get().getLanguage().equalsIgnoreCase("English")){
+            throw new NotAllowedLanguageException("It is not allowed to delete English language from database");
+        }
+        messageRepository.findAll().forEach(message -> {
+            if (message.getLanguage().getLanguage().equals(languageOptional.get().getLanguage())) {
+                messageRepository.deleteById(message.getId());
+            }
+        });
         languageRepository.deleteById(id);
     }
